Modified the token validation
Token validation was performed only on the local cache.
This was creating problem with external applications getting
the token from IAM bypassing Liferay, e.g. mobile applications.

The validation has been updated so if the token is not in cache
user information are retrieved from the user info endpoint.
This allows to validate all tokens released from IAM.